requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
python-dotenv
Flask-Mail
requests
Flask-WTF
email_validator
        